Added support for getting remote ip from trusted proxies
In the config you can set $config['TRUSTED_PROXIES'] to an array of
ip addresses which should be trusted to get the real ip from.
If enabled the the real ip will be extracted from the x-real-ip header.
The header to use can be set by $config['REAL_IP_HEADER'].
The real ip will then be shown in the audit log and in the syslog.